+++ src/Benchmarks/ReflectionHydrationBench.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace EventSauce\ObjectHydrator\Benchmarks;
 
 use EventSauce\ObjectHydrator\ObjectHydrator;
-use EventSauce\ObjectHydrator\HydrationDefinitionProviderUsingReflection;
+use EventSauce\ObjectHydrator\HydrationDefinitionProvider;
 
 class ReflectionHydrationBench extends HydrationBenchCase
 {
