+++ src/HydrationDefinitionProvider.php
@@ -4,15 +4,101 @@ declare(strict_types=1);
 
 namespace EventSauce\ObjectHydrator;
 
-use Throwable;
+use ReflectionClass;
+use ReflectionIntersectionType;
+use ReflectionMethod;
+use function count;
+use function is_a;
 
-/**
- * @internal
- */
-interface HydrationDefinitionProvider
+final class HydrationDefinitionProvider
 {
-    /**
-     * @throws Throwable
-     */
-    public function provideDefinition(string $className): ClassHydrationDefinition;
+    private DefaultCasterRepository $defaultCasterRepository;
+    private KeyFormatter $keyFormatter;
+
+    public function __construct(
+        DefaultCasterRepository $defaultCasterRepository = null,
+        KeyFormatter $keyFormatter = null,
+    ) {
+        $this->defaultCasterRepository = $defaultCasterRepository ?? DefaultCasterRepository::builtIn();
+        $this->keyFormatter = $keyFormatter ?? new KeyFormatterForSnakeCasing();
+    }
+
+    public function provideDefinition(string $className): ClassHydrationDefinition
+    {
+        $reflectionClass = new ReflectionClass($className);
+        $constructor = $this->resolveConstructor($reflectionClass);
+
+        /** @var PropertyHydrationDefinition[] $definitions */
+        $definitions = [];
+
+        $constructionStyle = $constructor instanceof ReflectionMethod ? $constructor->isConstructor(
+        ) ? 'new' : 'static' : 'new';
+        $constructorName = $constructionStyle === 'new' ? $className : $this->stringifyConstructor($constructor);
+        $parameters = $constructor instanceof ReflectionMethod ? $constructor->getParameters() : [];
+
+        foreach ($parameters as $parameter) {
+            $accessorName = $parameter->getName();
+            $key = $this->keyFormatter->propertyNameToKey($accessorName);
+            $parameterType = PropertyType::fromReflectionType($parameter->getType());
+            $firstTypeName = $parameterType->firstTypeName();
+            $keys = [$key => [$key]];
+
+            $attributes = $parameter->getAttributes();
+            $casters = [];
+
+            foreach ($attributes as $attribute) {
+                $attributeName = $attribute->getName();
+
+                if ($attributeName === MapFrom::class) {
+                    $keys = $attribute->newInstance()->keys;
+                }
+
+                if (is_a($attributeName, PropertyCaster::class, true)) {
+                    $casters[] = [$attributeName, $attribute->getArguments()];
+                }
+            }
+
+            if ($firstTypeName && count($casters) === 0 && $defaultCaster = $this->defaultCasterRepository->casterFor(
+                    $firstTypeName
+                )) {
+                $casters = [$defaultCaster];
+            }
+
+            $definitions[] = new PropertyHydrationDefinition(
+                $keys,
+                $accessorName,
+                $casters,
+                $parameterType,
+                $parameterType->canBeHydrated(),
+                $parameterType->isEnum(),
+                $parameterType->allowsNull(),
+                $firstTypeName,
+            );
+        }
+
+        return new ClassHydrationDefinition($constructorName, $constructionStyle, ...$definitions);
+    }
+
+    private function resolveConstructor(ReflectionClass $reflectionClass): ?ReflectionMethod
+    {
+        $methods = $reflectionClass->getMethods(ReflectionMethod::IS_STATIC | ReflectionMethod::IS_PUBLIC);
+
+        foreach ($methods as $method) {
+            $isConstructor = $method->getAttributes(Constructor::class);
+
+            if (count($isConstructor) !== 0) {
+                return $method;
+            }
+        }
+
+        return $reflectionClass->getConstructor();
+    }
+
+    private function stringifyConstructor(ReflectionMethod $constructor): string
+    {
+        $name = $constructor->getName();
+        $className = $constructor->getDeclaringClass()->getName();
+
+        return "$className::$name";
+    }
 }
