+++ src/ObjectHydratorTest.php
@@ -8,7 +8,7 @@ class ObjectHydratorTest extends ObjectHydratorTestCase
 {
     protected function createObjectHydrator(HydrationDefinitionProvider $definitionProvider = null): ObjectHydrator
     {
-        $definitionProvider ??= new HydrationDefinitionProviderUsingReflection(
+        $definitionProvider ??= new HydrationDefinitionProvider(
             keyFormatter: new KeyFormatterWithoutConversion()
         );
         return new ObjectHydrator($definitionProvider);
